Refactor AuthenticationManager into interface and implementation
